I add 2 songs for sale on same contract with different tokenID.
