+++ components/MintButton/index.js
@@ -1,0 +1,3 @@
+import MintButton from './MintButton'
+
+export default MintButton
