+++ lib/getDefaultProvider.js
@@ -4,7 +4,7 @@ const getDefaultProvider = (networkName, chainId) => {
   return chainId
     ? ethers.getDefaultProvider(
         networkName?.includes('mum')
-          ? 'https://rpc-mumbai.maticvigil.com'
+          ? 'https://matic-mumbai.chainstacklabs.com'
           : { chainId: parseInt(chainId.toString()), name: networkName }
       )
     : null
