+++ pages/index.tsx
@@ -20,14 +20,13 @@ export const getServerSideProps: GetStaticProps = async (context) => {
   )
   const provider = getDefaultProvider(chain.network, chainId);
   const contract = new ethers.Contract(contractAddress.toString(), abi, provider);
+
   // Get metadata renderer
   const songURI = await contract.songURI(1);
   const songURI2 = await contract.songURI(2);
   
   const metadataURI = ipfsImage(songURI)
   const metadataURI2 = ipfsImage(songURI2)
-  console.log("metadataURI", metadataURI)
-  console.log("metadataURI2", metadataURI2)
   const axios = require('axios').default;
   const {data: metadata} = await axios.get(metadataURI)
   const {data: metadata2} = await axios.get(metadataURI2)
@@ -35,8 +34,6 @@ export const getServerSideProps: GetStaticProps = async (context) => {
   const saleDetails = await contract.saleDetails();
   const erc721Drop = getDrop(contractAddress, metadata, saleDetails)
   const erc721Drop2 = getDrop(contractAddress, metadata2, saleDetails)
-  console.log("ERC721Drop", erc721Drop)
-  console.log("ERC721Drop2", erc721Drop2)
 
   return {
     props: { collection: erc721Drop, chainId: chain.id, collection2: erc721Drop2 },
